Don't check SAL_DETAIL_ENABLE_LOG_INFO, check SAL_LOG_INFO
SAL_DETAIL_ENABLE_LOG_INFO is always defined, as "true" or "false". It
is SAL_LOG_INFO that is defined or not, and can be used to avoid
unnecessary non-trivial code that the compiler might not be able to
optimise away.

Change-Id: I9903faab64b39fee86bff3b085cdce3614b0d921
(cherry picked from commit d23a1f4dabda939a216293c150c3679b0979d0bd)