Speed up in-process caching of OpenGL shader programs
It seems to be fairly CPU intensive to calculate the MD5 digest of
shader program source code. But we don't need to use that to look up a
corrresponding program in the run-time in-process cache anyway. The
shader names are unique, so it is enough to use that as key.

Change-Id: I8fd9f5f875be14a82cd53daf8a2ca72bfd23beb6