mailmerge: Make the SwMailMergeConfigItem instance long-living.
In order for the toolbar to work, we need the SwMailMergeConfigItem instance
in the view living even when the mail merge wizard is not around.

With this change, there's a small glitch that the SwMailMergeChildWindow
sometimes stays open even when we leave the wizard, but I plan to remove this
window / button anyway, so no real harm at the moment.

Change-Id: I2bf0c49504b9cf2f17f6fce7ab4cab7cb5be6426